add Behavior objects
I implemented the Behavior abstract class and two subclasses
SimpleBehavior and ComplexBehavior,which respresent a single set of motor
frequencies and a series of other Behaviors respectively. Also created an
Exception for invalid frequency values, a class to hold constants in the
Object package, and a stub for the TensegrityController class. The stub
currently just prints out the frequencies you sent to it.