Add numpy to venv, add Point code, fninsh others
I added numpy to the Venv to handle the 3D arrays in the Repertoire. I
also finished up preliminary versions of Behavior, Locomotor, and
Repertoire. That includes adding a MotorController class in Locomotor.py,
which will eventually be subclassed for different hardware.

I added a Point class, which holds a location <x, y, theta> and can be
added or subtrcted from another point.

Finally, I added a SaveLoad file, which will eventually house the methods
of saving and loading TensRobots.