Begin Main code
I'm beginning the code for the main class, which will provide an
interactive menu for using the tensegrity. Should be a pretty simple
setup. This is more of just a distraction/boondogle to procrastinate from
working on the image recognition...

I also added the ability to test the motors in a TensRobot