Switched to C++, using LIMBO
First, I apologize for the giant, messy push. This is all the new C++
code using LIMBO. I adapted the tracking software developed in France to
use our (extremely inferior) OpenCV puff ball tracking setup. Many
adjustments in the code were made to get the code to actually compile
and run without errors. As of this push, the code will generate 10
random samples, then run 40 iterations of the Bayesian Optimization
code. During each run the program will both display and record a video
of the tensegrity with a tracking overlay (center of tensegrity and its
heading). It will also output a text file of the location data
throughout each run.